Move send_message wrappers to public client
This move was made so get_product_trades can
return a generator like all other paginated
API endpoints.
